Implemented changes as requested in comments
