+++ frontend/src/components/Buttons/CreateNewLogButton.css
@@ -1,0 +1,4 @@
+.add-icon {
+    margin-right: 9px;
+    vertical-align: -6px;
+}
