+++ frontend/src/components/Buttons/HomeButton.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { useNavigate } from "react-router-dom";
+
+export default function HomeButton() {
+  const navigate = useNavigate();
+
+  const handleHomeClick = () => {
+    navigate("/dashboard");
+  };
+
+  return <button onClick={handleHomeClick}>Home</button>;
+}
