+++ frontend/src/components/Buttons/LogHistoryButton.css
@@ -1,0 +1,4 @@
+.history-icon {
+    margin-right: 9px;
+    vertical-align: -6px;
+}
