+++ frontend/src/components/Buttons/SignUpButton.css
@@ -1,0 +1,4 @@
+.sign-up-button-container {
+    display: flex;
+    gap: 8px;
+}
