+++ frontend/src/components/Navbar/Navbar.css
@@ -19,3 +19,14 @@
     margin-top: 10px;
     margin-right: 30px;
 }
+
+.sign-in-button-container {
+    display: flex;
+    gap: 9px;
+  }
+  
+  .login-icon {
+    margin-left: 8px;
+    font-size: 1.2rem;
+    vertical-align: middle;
+  }
