+++ frontend/src/pages/LogHistory.jsx
@@ -3,6 +3,7 @@ import Navbar from "../components/Navbar/Navbar";
 import SearchFilterSort from "../components/LogHistory/SearchFilterSort";
 import LogTable from "../components/LogHistory/LogTable";
 import Pagination from "../components/LogHistory/Pagination";
+import "./styles/LogHistory.css";
 
 const LogHistory = () => {
   const initialLogs = [
