+++ frontend/src/pages/styles/Dashboard.css
@@ -18,8 +18,3 @@
     flex-direction: column;
     gap: 10px;
 }
-
-.add-icon {
-    margin-right: 9px;
-    vertical-align: -6px;
-}
