+++ frontend/src/pages/styles/Homepage.css
@@ -14,8 +14,3 @@
     width: 450px;
     margin-top: -1rem;
 }
-
-.sign-up-button-container {
-    display: flex;
-    gap: 8px;
-}
