+++ frontend/src/pages/styles/LogHistory.css
@@ -1,0 +1,9 @@
+.log-history-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+.container {
+    width: 80%;
+}
