+++ frontend/src/pages/styles/NewLog.css
